# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
      "ActivitiesListView": {
        "properties": {
          "activities": {
            "items": {
              "$ref": "#/components/schemas/ActivityItem"
            },
            "title": "Activities",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "activities",
          "total"
        ],
        "title": "ActivitiesListView",
        "type": "object"
      },
...
      "Body_import_garmin_data": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          }
        },
        "title": "Body_import_garmin_data",
        "type": "object"
      },
...
      "GarminImportResponse": {
        "properties": {
          "imported_dates": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Imported Dates",
            "type": "array"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "GarminImportResponse",
        "type": "object"
      },
      "GarminSyncResponse": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "synced_days": {
            "default": 0,
            "title": "Synced Days",
            "type": "integer"
          }
        },
        "required": [
          "ok"
        ],
        "title": "GarminSyncResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NightsListView": {
        "properties": {
          "nights": {
            "items": {
              "$ref": "#/components/schemas/NightListItem"
            },
            "title": "Nights",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "nights",
          "total"
        ],
        "title": "NightsListView",
        "type": "object"
      },
...
      "NotFoundBody": {
        "description": "404 — the module behind the endpoint is switched off, or nothing is there.",
        "properties": {
          "error": {
            "enum": [
              "module_disabled",
              "not_found"
            ],
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "NotFoundBody",
        "type": "object"
      },
...
      "RecoveryBar": {
        "properties": {
          "hi": {
            "title": "Hi",
            "type": "number"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "lo": {
            "title": "Lo",
            "type": "number"
          },
          "max": {
            "title": "Max",
            "type": "number"
          },
          "min": {
            "title": "Min",
            "type": "number"
          },
          "tone": {
            "default": "",
            "title": "Tone",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "key",
          "min",
          "max",
          "lo",
          "hi"
        ],
        "title": "RecoveryBar",
        "type": "object"
      },
...
      "RecoveryView": {
        "properties": {
          "bars": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RecoveryBar"
            },
            "title": "Bars",
            "type": "array"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "days": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RecoveryDayItem"
            },
            "title": "Days",
            "type": "array"
          },
          "headline": {
            "$ref": "#/components/schemas/RecoveryHeadline"
          },
          "is_configured": {
            "title": "Is Configured",
            "type": "boolean"
          },
          "is_today": {
            "title": "Is Today",
            "type": "boolean"
          },
          "last_sync": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync"
          },
          "night": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecoveryNightPreview"
              },
              {
                "type": "null"
              }
            ]
          },
          "norms": {
            "additionalProperties": {
              "$ref": "#/components/schemas/RecoveryNorm"
            },
            "title": "Norms",
            "type": "object"
          },
          "today_date": {
            "format": "date",
            "title": "Today Date",
            "type": "string"
          }
        },
        "required": [
          "date",
          "today_date",
          "is_today",
          "is_configured",
          "headline",
          "norms"
        ],
        "title": "RecoveryView",
        "type": "object"
      },
...
      "SleepNightView": {
        "properties": {
          "awake_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awake Count"
          },
          "awake_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awake Seconds"
          },
          "bb_change": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bb Change"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "deep_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deep Seconds"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "heart_rate": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/IntradaySeriesPoint"
            },
            "title": "Heart Rate",
            "type": "array"
          },
          "hrv": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/IntradaySeriesPoint"
            },
            "title": "Hrv",
            "type": "array"
          },
          "is_today": {
            "title": "Is Today",
            "type": "boolean"
          },
          "light_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Light Seconds"
          },
          "movement": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/IntradaySeriesPoint"
            },
            "title": "Movement",
            "type": "array"
          },
          "next_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Date"
          },
          "prev_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev Date"
          },
          "rem_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rem Seconds"
          },
          "respiration": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/IntradaySeriesPoint"
            },
            "title": "Respiration",
            "type": "array"
          },
          "restless_moments": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Restless Moments"
          },
          "rhr": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rhr"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "spo2_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spo2 Min"
          },
          "stages_minutes": {
            "additionalProperties": {
              "type": "integer"
            },
            "default": {},
            "title": "Stages Minutes",
            "type": "object"
          },
          "stages_series": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SleepStageSegment"
            },
            "title": "Stages Series",
            "type": "array"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "today_date": {
            "format": "date",
            "title": "Today Date",
            "type": "string"
          }
        },
        "required": [
          "date",
          "today_date",
          "is_today"
        ],
        "title": "SleepNightView",
        "type": "object"
      },
...
      "UnauthenticatedBody": {
        "description": "401 — no valid session cookie.",
        "properties": {
          "error": {
            "const": "unauthenticated",
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "UnauthenticatedBody",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Vitals API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/recovery": {
      "get": {
        "description": "The Recovery screen: sleep score, HRV, resting HR, body battery, night preview,\nand 14-day recovery grid.",
        "operationId": "read_recovery_overview",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundBody"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Read Recovery Overview"
      }
    },
    "/api/v1/recovery/activities": {
      "get": {
        "description": "List recorded sport activities, newest first.",
        "operationId": "read_activities_list",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivitiesListView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Activities List"
      }
    },
    "/api/v1/recovery/import": {
      "post": {
        "description": "Ingest a Health Auto Export JSON dump.",
        "operationId": "import_garmin_data",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_garmin_data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GarminImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Garmin Data"
      }
    },
    "/api/v1/recovery/nights": {
      "get": {
        "description": "List recorded sleep sessions, newest first.",
        "operationId": "read_nights_list",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 60,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NightsListView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Nights List"
      }
    },
    "/api/v1/recovery/sleep/{on_date}": {
      "get": {
        "description": "One night in detail: hypnogram, minute-level curves, left/right night navigation.",
        "operationId": "read_sleep_night_detail",
        "parameters": [
          {
            "in": "path",
            "name": "on_date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "On Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SleepNightView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Sleep Night Detail"
      }
    },
    "/api/v1/recovery/sync": {
      "post": {
        "description": "Trigger an on-demand sync of Garmin metrics.",
        "operationId": "sync_garmin_now",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GarminSyncResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundBody"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Sync Garmin Now"
      }
    },
...
  }
}